Deduct book quantity based on sales detail transaction

+++ dto/book-dto.go
@@ -3,7 +3,7 @@ package dto
 type BookRequest struct {
 	Title  string `json:"title" form:"title"`
 	Author string `json:"author" form:"author"`
-	Price  int `json:"price" form:"price"`
+	Price  float64 `json:"price" form:"price"`
 	Stock  int `json:"stock" form:"stock"`
 	CategoryId  int `json:"category" form:"category"`
 }
